predict_engine_research/pyproject.toml
[project]
name = "predict-engine-research"
version = "0.1.0"
description = "CatBoost regression research project"
requires-python = ">=3.12"
dependencies = [
    "catboost>=1.2.10",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "wandb>=0.25.1",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/envs", "src/models", "src/runtime", "src/tracking"]
